restructure to separate components and add custom css
