+++ assets/custom.css
@@ -1,0 +1,7 @@
+div {
+}
+
+.mycol {
+    border: 1px solid blue;
+}
+
