+++ components/MultiStepProgressBar.py
@@ -1,0 +1,10 @@
+import solara
+
+@solara.component_vue("MultiStepProgressBar.vue")
+def MultiStepProgressBar(
+    steps = None,
+    currentStep = None,
+    currentStepProgress = None,
+    height = None,
+    ): pass
+
