+++ components/TableDisplay.py
@@ -1,0 +1,10 @@
+import solara
+
+@solara.component
+def TableDisplay(*args, **kwargs):
+    """
+    Thin wrapper around solara.DataFrame
+    """
+    return solara.DataFrame(*args, **kwargs)
+
+
